pyproject.toml
[tool.poetry]
name = "proxygen-cli"
version = "0.1.0"
description = "CLI for interacting with NHSD APIM's proxygen service"
authors = ["Ben Strutt <ben.strutt1@nhs.net>"]
readme = "README.md"
packages = [{include = "proxygen_cli"}]
repository = "https://github.com/NHSDigital/proxygen-cli"

[tool.poetry.scripts]
proxygen = 'proxygen_cli.main:main'

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.10.2"
click = "^8.1.3"
pyjwt = "^2.6.0"
requests = "^2.28.1"
lxml = "^4.9.1"
cryptography = "^38.0.3"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
